backend/requirements.txt
alembic==1.12.0
asgiref==3.7.2
blinker==1.6.2
braintree==4.23.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
distlib==0.3.7
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
filelock==3.12.4
Flask==2.3.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
fonttools==4.42.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.0
packaging==23.1
Pillow==10.0.1
pipenv==2023.10.20
platformdirs==3.11.0
psycopg2-binary==2.9.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
virtualenv==20.24.5
Werkzeug==2.3.7